# swagger/api-docs.yaml
openapi: 3.0.0
paths:
...
  /study-goals:
    get:
      operationId: StudyGoalController_findAll
      parameters:
        - name: activeOnly
          required: true
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: ''
      tags: &ref_3
        - StudyGoal
    post:
      operationId: StudyGoalController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateStudyGoalDto'
      responses:
        '201':
          description: ''
      tags: *ref_3
  /study-goals/category/{categoryId}:
    get:
      operationId: StudyGoalController_findByCategory
      parameters:
        - name: categoryId
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags: *ref_3
...
  /categories:
    get:
      operationId: CategoryController_findAll
      parameters:
        - name: activeOnly
          required: true
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: ''
      tags: &ref_4
        - Category
    post:
      operationId: CategoryController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCategoryDto'
      responses:
        '201':
          description: ''
      tags: *ref_4
  /categories/{id}:
    get:
      operationId: CategoryController_findById
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags: *ref_4
    put:
      operationId: CategoryController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags: *ref_4
    delete:
      operationId: CategoryController_delete
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags: *ref_4
...
info:
  title: TestFlow API
  description: API para la aplicación TestFlow
  version: '1.0'
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    CreateStudyGoalDto:
      type: object
      properties: {}
    CreateCategoryDto:
      type: object
      properties: {}
...